feat: useStickyMode -> useStickyOperation변경
