+++ packages/design-systems/src/Sticky/hooks/index.ts
@@ -2,3 +2,4 @@ export { default as usePositionCalculators } from "./usePositionCalculators";
 export { default as useStatusUpdaters } from "./useStatusUpdaters";
 export { default as useStyles } from "./useStyles";
 export { default as useStickyMode } from "./useStickyMode";
+export { default as useStickyOperation } from "./useStickyOperation";
