+++ packages/design-systems/src/Sticky/hooks/useStickyMode/index.ts
@@ -1,2 +1,2 @@
-export * from "./useStickyMode";
 export { default } from "./useStickyMode";
+export { default as stickyRenderMode } from "./stickyRenderMode";
